usb-creator.sh: Fixup checks that config_file exists
The current checks would pass if the value passed for --config was a
relative path that exists but the script would then proceed to fail
since it then 'cd' to another location before sourcing the
config_file. Add to the checking code to convert relative paths used
for --config to be an abs path to avoid this. Also provide more
information in the error message given if it fails to find the
config_file to assist the user in passing a valid value to --config.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>