functions.sh: Fix issues when block device has no part table
When using NBD, or possibly when a never before used block device, is
used as the installer media the installer will fail while attempting
to create partitions on the device. Check for this condition and use
parted 'mklabel' to initialize the partition table.

We also quiet errors that parted might throw when no partition table
exists and the installer is probing the device for existing
partitions.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>