misc: allow config to skip confirm_install
As things are now the installer is always somewhat interactive, it
will always prompt the user if it should continue the install after
displaying the installation configuration. If for some reason someone
wants to script this process they would end up having to use some sort
of expect script to complete the interactive parts. Add a
configuration which can allow the installation to proceed without the
user being prompted for confirmation.

The default behavior is to continue to prompt the user.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>