puppet: add support for puppet to the installer
This is the first step in adding support for puppet to our simple
overc installer. Here we provide a simple example puppet configuration
which will set the hostname and add a 'guest' user to the
pod-essential install. We also provide a configuration to define the
location of the puppet files. Lastly we provide updated scripts to
allow the configs to be copied to the install media and for settings
to be applied at the end of the installation script.

Although this initial implementation is quite simple we can
potentially do quite a bit of configuration of the installation by
chaining puppet calls and using chrooted environments for the pods
that are installed. Many of the steps we have in the pod-install.sh
script could already be removed and replaced by puppet configurations.

A section will be added to meta-overc/README.install to cover using
puppet. To start puppet will remain an optional feature of the
installer.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>