pod-install: able to set container to control serial console
Currently serial device (most likely to be ttyS0) are available for
essential and for all containers in overc.  Also each container will
spawn agetty on serial device which causes the fighting of serial
console between essential and container.  It would be good to make sure
that only essential or a container is in-charge of serial console.  And
we prefer this ability during installation process.

pod-install allows to pass in container name for controlling
serial console.  pod-installer will enable serial-getty@XXXX.service
service for the provided container and mask out this service for all
other container and essential.

Signed-off-by: Vu Tran <vu.tran@windriver.com>