installer: allow board specific functions files
The structure of the installer puts common routines into a file
called 'functions.sh' and a specific installer includes it, and
adds custom routines as necessary (pod-install.sh and usb-creator.sh
are two examples).

This means that common routines must be provided in the central
'functions.sh' file, with the ability to handle special cases
relegated to if statements and special processing.

To avoid this sort of conditional statement, this commit allows
board specific functions files to be part of the installer 'sbin'
directory, or in the ~/.overc/ directory.

They are of the format: functions-$BOARD_NAME.sh, where BOARD_NAME
is defined in the various config.sh files.

These board specific functions are available to the host USB
creator, as well as the on-target installer(s).

We also add to the common 'functions.sh' file a new routine
called "override_function". By calling this, a board specific
function file can replace/override a common routine from
functions.sh. This is particularly useful for bootloader and
board specific routines. See functions-fsl-ls10xx.sh and
functions-xilinx-zynq.sh for samples of this technique.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>