grub: fix install in host/loopback configuration
When running the installer against a loopback device (versus on
target) grub can't probe the proper modules to access the boot
device later.

As such, we detect that we are on a loopback device, and run grub
install with modules specified.

Also, we take the device mapper out of the picture and use partprobe
versus kpartx

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>