overc-installer: fix the bug of cannot install initramfs
When build a bootable image in wr-core build directory,
there will be several initramfs files in tmp/deploy/images/
directory, one is a soft link such as cube-server-xilinx-zynq.tar.bz2,
and one is with timestamp in the its name, such as
cube-server-xilinx-zynq-20151027033017.rootfs.tar.bz2, so in this case,
only one of them is isntalled.

BTW, the config file specified one is the first choice.

Signed-off-by: fli <fupan.li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>