Allow command-line specification of disk size
On boards that boot from removable media, the size of the disk
being used may not be known until the installer is actually run.
So it makes sense to allow it as a command-line option in addition
to specifying it as a board-specific option.

Also set the default size to 1GB less than a common removable
media size.  This is because many commercial disks do not actually
provide the advertized size to the user.  For example, a specific
Kingston disk labelled 8GB actually contained 7969177600 bytes,
or 7.6 * 1000 * 1024 * 1024.

(One might assume that the extra space is used for storage by
the internal processor and/or for replacement sectors.)

Signed-off-by: T.O. Radzy Radzykewycz <radzy@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>