Resolve initrd conflict
The copy command must either copy a single file to a destination
file or directory, or copy a group of files to a directory.  It is
not allowed for the copy command to copy multiple files to a file.

The copy command is inside a loop, but was wrongly selecting the
files that the loop iterated over instead of the index variable.

Signed-off-by: T.O. Radzy Radzykewycz <radzy@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>