installer: use variables instead of container_dir
The installer runs in a more restricted environment when it runs
natively (versus being an image installer), where only required
elements are present.

While this is nice, it means that the behaviour is different than
live image creation.

To unify the two methods, we generate a config file for the installer
(with all the attributes of the bootable USB creation) and use that
file when running on target.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>