Move mount point to temp dir
Create a temporary directory to mount the filesystem on.  This is
a necessary requirement for the installers to be reentrant.
Some users may wish to run multiple installers simultaneously,
for example to create image files for two different boards or to
create an installer image file and a bootable image file, or to
create bootable image files for different sized storage devices.

Signed-off-by: T.O. Radzy Radzykewycz <radzy@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>