installer: don't remove the mount dir when it is still needed
We are removing the temporary mount directory, and then trying to use
it again for a btrfs snapshot. This won't work, unless we want to
re-create it before that mount. It is easier to simply remove it once
when we are done using it, so we move that cleanup to the bottom of
the installer.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>