config-usb*: cleanup PREREQ file checks
config-usb-cube.sh sources config-usb.sh right away to allow default
values to be set which config-usb-cube.sh can then overwrite. This
works as expected but config-usb.sh sets PREREQ_FILES which the
installer uses to check for the existence of files. So even though the
defaults are overwritten, if the default files don't exist the
installer will fail to run.

Allow PREREQ_FILES to be recalculated after we overwrite variables by
moving its calculation into a function that we can call from
config-usb-cube.sh. Since we have this PREREQ_FILES check we might as
well expand the list of files that is checked as well.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>