overc-installer: remove the consumed remount of root partition from fstab
Since the remount of root partation is a time consuming step,
it's better to remove it from fstab and resort to the boot
parameter passed to kernel.

Another, the kernel api filesystem's mount will be take cared by
systemd's static unit and it safe to remove them from fstab to
reduce the boottime.

By removing these remount, the boottime will be reduced by
20 seconds.

Signed-off-by: fupan li <fupan.li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>