overc-installer: Add the fdisk partition layout file input function
Some bsps need deploy the "u-boot" on the SD card beginning sectors.
and the u-boot environment parameters are saved in the SD card beginning
1MB. The "fdisk" tool creates the first partition from the 2048 sectors
(1MB) by default.

If user want to save the environment parameters in the u-boot, then
it will destroy the SD card file system's superblocks in the 1MB.

So reserve the beginning 2MB for the bootloader and environment
parameters.

Add the --partition_layout parameter to supply other fdisk partition
layout file from the command line. If command line does not include this
parameter, then the install_templates configure will be used as default.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>