overc-cctl: add support for bridged or vswitch networking
In order to allow the installer to make use of OpenVSwitch networking,
which may be built into the OverC, containers and -essential we need
to add support for a switch to tell us which networking to configure
in the lxc config files for containers.

We add a 'b' option to add_container to allow the installer to
configure the legacy bridge networking. During upgrade the previous
configuration will be polled and used.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>