network-prime: Add the ability to configure the networking prime
The new overc-conftools package which has been added to the -essential
image in meta-overc allows for system configuration during the first
and subsequent boots of a system built on the OverC framework. This
allows for configuring the system beyond what is already available at
package, dist, image and installer creation. The configurations
included in overc-conftools can be fine tuned by the installer, which
is what we are doing here with setting the networking prime and static
IPs.

Here we add support for a new attribute for HDINSTALL_CONTAINERS,
similar to the vty attribute, which allows us to configure the
networking prime as well as specify statc IP offsets for
containers. The attribute is named "net" and can be used as follows:

HDINSTALL_CONTAINERS="${ARTIFACTS_DIR}/cube-dom0-genericx86-64.tar.bz2:vty=2:net=3 \
                      ${ARTIFACTS_DIR}/cube-server-genericx86-64.tar.bz2:vty=3 \
                      ${ARTIFACTS_DIR}/cube-domE-genericx86-64.tar.bz2:vty=4:net=1"

In the above domE will become the networking prime. dom0 will get an
offset of "3" and therefor get 192.168.0.3 as its static
IP. cube-server will get a dynamic IP.

If you do not put a 'net=1' attribute on any container the installer
will default to use bridged networking. This requires that you
complete your build with the 'overc-bridge-networking' DISTRO_FEATURE.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>