cubeit-installer: Add overc.network.link function as a condition
The overc.network.link funtion can be used from an install template by
setting NETWORK_DEVICE=all and optionally passing a string of devices
for the network prime with
  NETWORK_DEVICE_CLASSES=[one or more devices with a wild card].

In order to generate images which will properly boot on hardware with
various network devices names it is desireable for some configurations
to own all the physical network devices specified by wild cards.

The additional specification in the lxc config file is:

overc.network.link = all
overc.network.linkdevs = eth* wl* en*

When the overc.network.link is set to all, the fixed or wild card
network devices can all be assigned to the container when the puppet
scripts initialize the first time.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>