cubeit-installer/cubename: allow file prefix
The cubename helper script had a placeholder for a 'prefix' component
to the cube filename, implement the placeholder. To have the prefix
handled correctly we require changes to the cubeit-installer and the
helper functions.sh. The workflow when building an installer remains
the same, if you do use a prefix for your cube filenames you can now
specify one and the installer will handle this.

Using the cubeit-installer this requires setting a value for
'CONTAINER_PREFIX' in your config files. If an empty string is
supplied for CONTAINER_PREFIX or it is simply not set the previous
behaviour persists.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>