cubeit-installer: cleanup container tarballs
As we install each container we create a copy of the container tarball
to ${TMPMNT}/tmp. Depending on how many containers we install this can
cause the partition to run out of space. Since we don't use these
tarballs after the container is installed we can do some cleanup after
we install each container to avoid space issues.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>