cubeit-installer: respect prefix
Work done on commit 622ca316365d250c517705312df1aaa20b32bbe6
[cubeit-installer/cubename: allow file prefix] was tested with and
without prefixes but prior to the pull request being sent some changes
for shared filesystem were merge. When testing the rebase only the
non-prefix case was tested and it was not noticed that several new
calls to the cubename script were added.

Adding the passing of the prefix to these new call sites.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>