cubeit: support a more extensive target config
Currently '--target-config' is only used for live images, in the case
of using non-live (2 stage install) we generated a limited target
config. In order to support the work done for the network prime we
need to make additions to this config file generation, adding
NETWORK_DEVICE and NETWORK_DEVICE_CLASSES. This exercise must be
repeated whenever we add new functionality that may depend on
variables being set in the config. Additionally we would need to
extend the configuration generation to cover 100% of the variables
that can be set in the installer configuration to unlock all the
capabilities of the installer.

The above has already demonstrated that the config generation is a bit
of a maintenance issue and needs to be revisited. With this set of
changes we can continue to build on the config generation and fall-back
to its use but we now reuse the '--target-config' parameter to allow a
user to pass a manually created config that will be copied to the
installer media (with some convenience to allow for the --config file
to be reused.)

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>