overc-installer: move the overlayfs mount entry from essential's fstab to container
If the domain container's overlayfs's mount is done in essential
system, the runtime container's rollback will break this mount,
which will make the start container failed after rollback without
remount this mount entry.

So the best place to mount the container's overlayfs is in its
mount file of /var/lib/lxc/<container>/fstab.

Signed-off-by: fli <fupan.li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>