cubeit-installer: don't rely on host umask
As it is we are at the mercy of the umask that is set for the user
that is running the installer, whether doing a live or 2-stage
install. Worst case this results in busted container rootfs where only
root has permissions to do anything; where the umask is 0077 and thus
the rootfs created by mkdir will be 'drwx------'. In order to be
consistent for all calls to tools like mkdir and others with use umask
we set a umask at the top of the installer. We will handle any
exceptions as needed, though at this time there does not appear to be
any.

We set the umask to 0022 as this represents the mask we have been
using most.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>