Improve cold boot time: Remove hard sleep from rootdelay
On modern  hardware the  rootdelay is not  needed.  The  rootwait will
wait for any kind of slow responding  USB device that takes a while to
come back from a bus reset.

Also the default grub timeout goes from 5 seconds to 1 second, which
means you can still interrupt boot sequence if absolutely needed, but
the general case of cold booting is faster.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>