overc-installer: delay cleanup the snapshots when rollback itself
When container do rollback itself, such as rollback dom0 in
dom0, it shouldn't do clean at present, instead it should
delay the cleanup to the next boot up, otherwise, dom0
will crash and cannot do any commands.

Add an relaunch option "-l" to rollback. If do rollback dom0
in itself and pass "-l" also, it will reboot the system immediately
once rollback succesully, otherwise, it will print an waring
message to tell users to reboot the system as soon as possible.

Another, since there is already an option "-r" for rollback
to remove the unused snapshot of a container, and the codes
added by:

commit 74a17c9f149aed5 <overc-cctl: Purge stray snapshots after a rollback operation>
are redundant, thus removed them, and users can pass "-r" to
rollback command to remove the stray snapshot.

Signed-off-by: fli <fupan.li@windriver.com>