config: add default configure file
Currently the configuration items are to be defined by users, somehow
we should provide the configureation item list and explanations for
user to choose. So we add a default configuration file for user's
information, and user just have to define item that they want to
override in their own config file.

Signed-off-by: Feng Mu <Feng.Mu@windriver.com>