cubeit-installer: Allow user to append kernel cmdline params
The installer uses a fixed grub configuration file at the moment, this
change allows some customization by the user by allowing them to
append to the kernel cmdline params. With this change they can, for
example, to sredirect the console output, or disable lapic.

Since I had to dig through the code to remind myself about how some of
the variables we set in the installer configuration I decided to start
documenting some of the more common variables to help myself and
others recall what these variables do. In order not to have thise
README show up in the github landing page a basic overview README was
added as well.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>