installer: add interactive configuration process
Currently user must define a configuration file to lead the
installation process. In order to reduce the learning complexity of the
configuration items, we introduce this interactive process to guide the
user to genarate a configuration file and use that during the following
installation process.

This interactive process is triggered if no configuration file is
provided through the commandline or if interactive parameter
--interactive or -I is specifically provided through commandline.

The genarated configuration file can be reused later by passing through
the --config parameter.

Signed-off-by: Feng Mu <Feng.Mu@windriver.com>