cubeit-installer: Use the INSTALL_KERNEL if specified
The INSTALL_KERNEL variable was used in a prior version of the cubeit
script.  Restore the functionality to use the INSTALL_KERNEL because
sometimes the provided boot volume will not have the right kernel or
no kernel at all.

This allows for iterative development using the
tmp/deploy/image/*Image file directly and respining full images
without rebuilding the rootfs.

The initrd and *Image* files were all getting copied more than they
should.  We need a backup *Image* file for the fallback boot but that
is it.

The function to directly use a kernel module tarball via
INSTALL_MODULES has also been restored.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>