cubeit-installer: smart installation for EFI boot artifacts
No matter how UEFI secure boot is configured, the efi boot artifacts are
always well-organized in /EFI/BOOT/ in ESP or fat filesystem. Basing on
the fact that the essential rootfs contains them in /boot/efi/EFI/BOOT/,
the efi boot installation is thus simplified as
"cp -a $rootfs/efi/EFI $esp/", where $rootfs is pointing to ./boot/
and $esp is pointing to ./mnt/.

In addition, introduce a new variable "INSTALL_GRUBEFI_CFG" to allow
specifying it externally in order to overwrite the default one.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>