cubeit-installer: use 'dos' if 'mbr' disklabel type not available
The 'mbr' disklabel type is an alias to the 'dos' disklable type and
was only introduced since util-linux v2.26 (see commit 4a79a8f1 on
git://git.kernel.org/pub/scm/utils/util-linux/util-linux.git). This
means that the 'mbr' type is not available in all fdisk version that
support the '-t' disklabel option.

In order to allow the installer to work seamlessly with a broad range
of fdisk versions we use the output of fdisk help to tailor which
options we need to set and which values we pass for these options. We
add logic to easily query supported options and to fix installer
creation issues with versions of fdisk which don't understand the
'mbr' disklabel type, we pass 'dos' instead.

NOTE that at this time 'mbr' and 'dos' types are synonymous but this
change will make distinguishing between the two easy should this ever
change and the need ever arise to deal with the disklabel differently.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>