github-fetcher.py: initial version of a simple github fetcher
This is a python module/stand-alone tool to fetch partial repository
contents from github via the github API (v3). To create a clone of a
repo it is always recommended to use git but if you only need one file
or the contents of a single directory then this tool can fetch this.

Other python scripts can include this module in order to make use of
the contained classes (only Fetcher at the moment). This script can
alternatively be run as a stand-alone tool. For example to fetch a
single file:

github-fetcher.py OverC meta-overc README /tmp

Will fetch the README file from OverC/meta-overc and copy it to /tmp
on the local machine. You may also fetch a directory:

github-fetcher.py OverC meta-overc conf /tmp

NOTE, however, that if you do not enable recursive fetch you will only
fetch the directory metadata. To fetch the directory and its contents
use:

github-fetcher.py -r OverC meta-overc README /tmp

This has been written to be python3 safe but a few adjustments
need to be made to make it work with python3, so for now only
python2 is supported.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>