cubeit-installer: make EFI boot robust
If EFI secure boot is enabled, the linux command can load the kernel image
but never verify it. So the post processing needs to replace it with the
chainloader command which supports to call shim protocol to verify the
kernel image.

In addition, the initrd command becomes not functional if the linux
command was not called ever.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>