cubeit: add 'dir' install type
This implements the previously stubbed out 'dir' install type. This is
similar to the installer + payload (bootable USB/on target installer)
type but only provides the payload part. This is useful to allow
installations over PXE and/or NFS or where a fixed kernel and
initramfs are used to bringup the target (factory reset) and the
payload can be accessed over NFS or removable media.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>