assert_return: fixup usage
The behavior of assert_return() is to change any non-zero return code
into a return code of "1". Calling assert_return() does not cause a
return in the calling function. Since assert_return() is itself a
function any calls to return simply return from assert_return().

Since all callers of assert_return() never check its return value
these calls are completely useless. Judging by the call sites it can
be assumed that there was an expectation that assert_return() would
cause a return at the call site. We therefor drop the use of
assert_return() and instead use '||' to validate return codes,
executing the code on the right of the '||' only when the code on the
left is evaluated as non-zero. In these cases returning 1 to implement
the intended behavior of assert_return().

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>