cubeit: split up custom_install_rules()
This function modifies media to

1) make the media bootable by copying/installing grub, a kernel, an
initramfs and a rootfs

2) have a copy the
  a) installer (static) scripts and configs
  b) installer (dynamic) data, such as scripts, rootfs, kernel and
     container images.

In this commit we keep all this functionality intact such that the
output remains identical, however, by putting (2.a) and (2.b) into
their own functions we make these distinct operations more
apparent. This also sets us up to be able to create an install bundle
which only contains the installer and installer while reusing these
new functions.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>