cubeit: copy LOCAL_POST_FUNCTION_DEFS directoy contents to media
If a LOCAL_POST_FUNCTION_DEFS is specified it currently isn't copied
to the installer media thus breaking the use of this mechanism for on
target install. Ensure the LOCAL_POST_FUNCTION_DEFS are copied along
with any other scripts found in the same directory, allowing the
script to be logically split into more than a single file.

We ammend the README.variables to make it clear that the entire
contents of the directory where this file resides will be copied,
hopefully avoiding any surprises.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>