cubeit-installer: drop priviledge console flag, and fix console attribute processing
Before the OCI support we would declare cube-desktop the console container if
one was not specified.

It so happens, that if a console is specified we were generating a pflask/cube
configuration since runc managecd containers do not properly take over a tty.

The end result of this was that if you did not specify dom0 as a console container,
and were installing cube-destkop it was incorrect flipped to be a pflask
container.

We fix this by specifying dom0 as the default console container, and also by
allowing an OCI container to generate a config if it is tagged as the console
device .. but note, that config won't do anything until tty handling chnages.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>