chvt: make chvt service OCI compliant
When a vty is tagged as active, we want it to be the default
login prompt presented to the user. The chvt service was out
of sync with the switch to runc, and also with recent systemd
changes.

This commit removes the lxc assumptions, changes the .service
to use cube-cmd and fixes the symlinks for a modern systemd.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>