cubeit-installer: Add another call to partx
A corner case was found where the partition table of a loop device
existed but was stale with respect to /sys and udev.  In this case the
kernel partition map must be dropped with the partx -d before trying
to perform an add, else the add will fail with an error that the
partition mapping already exists when in fact it is wrong.

Testing of all the known corner cases and default operation where the
loop device is first initialized show that it is safe to always drop
the partition map before performing the add with partx.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>