overc-installer: move chvt.service from foreground container to dom0
chvt.serice will use oci command to active the foreground vty,
but if the foreground container was cube-server or any other
containers, to run oci command will active ACL check and ask for
ID/passwd input, which is not possible to input them in a systemd
service. Thus it's better to put this service into dom0.

Signed-off-by: Fupan Li <fupan.li@windriver.com>