overc-ctl: allow non-tracked container to still be removed from disk
We sometimes end up with an unbalanced track/untrack pair. In that
scenario, we still want to remove the container from disk.

We also fix an incorrect test for a running container (to match the
rest of the calls in the tool), and fix some minor grammar in the
error messages.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>