sbin/functions.sh: use consistent partition names
Attempting to boot the live+payload installer will fail since the
stage1 install will create OVERCBOOT and OVERCINSTROOTFS on the
installer media yet grub will be configured to use INSTBOOT and
INSTROOTFS. Making the boot fail out of the box.

Make the naming consistent with the labels we have in the other
installer scripts.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>