cubeit: Add GPT support for 'installer' images
HW with UEFI is backwards compatible with MBR but GPT lacks many of
the limitations of MBR (such as unlimited partitions). We aren't
currently pushing the limits of MBR so we continue to use MBR by
default. But we make it possible to build 'installer' images
(live+payload) which use GPT.

The partition table type can be controlled via INSTALL_USE_GPT
(leaving it unset will build using MBR). This does not affect things
when you perform a 'dir' or 'image' install type.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>