cubeit: fixup installer (live+payload) with EFI
There have been changes in the way bitbake creates images since the
introduction of WIC images in Yocto, changes the installer is yet to
adapt to. In addition the use of cube-essential as the 'live' rootfs
is no longer viable as we have made efforts to shrink this image
resulting in it lacking many of the tools the installer needs to run.

These changes coincide with the introduction of the 'cube-install'
image in meta-overc, an image which has the necessary tools to run the
installer and can be used as a cube or rootfs. This solves the issue
of missing tooling.

The Yocto changes brought about for WIC images mean that
grub-efi.bbclass is not being 'executed'. We can therefor no longer
share grub (efi) artifacts from the rootfs and instead need to perform
a proper grub efi installation to populate the EFI System Partition
(ESP). The approach used here means that we are now consistent for
both EFI and non-EFI cases and frees us from depending on what the
image contains.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>