cubeit: enable the use of GPT
UEFI systems can understand MBR as part of the required backwards
compatibility but it is usually a good idea to use GPT when
possible. Using GPT for example has benefits such as exposing the
firmware in the EFI System Partition (ESP) instead of buried in the
MBR, and unlimited partitions.

In order to support GPT we switch out 'fdisk' use for 'parted', which
has GPT support built in. (fdisk can support GPT but only with the
help of other tools). As a side effect of using parted the 'scripting'
of the partitioning becomes much more readable.

GPT expects the ESP to be FAT32 so we force this while creating the
vfat filesystem.

The INSTALL_USE_GPT variable, which was setup for live+payload
installs, is reused here is used to trigger the use of GPT. By default
MBR is used.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>