cubeit/functions: work around multiple available grub targets
If the essential's grub is built with multiple targets enabled we
would fail to properly install grub with the error:

  /usr/sbin/grub-install: error: install device isn't specified.
  /bin/bash: line 1: x86_64-efi: command not found

This was a result of the multiple targets being passed --target and
thus the second one showing up as an option to grub-install.

Here we recongnize when there are multiple targets available and only
use the first one found or the one defined by the value of
GRUB_TARGET. If we do find multiple targets or if GRUB_TARGET doesn't
match a found target we print out helpful messages on how to correct
things, and fallback to one of the available targets.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>