cubeit: implement the 'mounts' container attribute
It is sometimes desirable to be able to configure mounts before the
first boot of a device so implement a 'mounts' container attribute to
allow this to be configured.

The syntax matches 'c3 cfg mount' and makes use of this functionality
directly. Since the mount is defined using a path (/blah/this/that)
this new attribute exposed a flaw in how we displayed the container
tarball names in the "Install Summary" so we fixed this by ensuring we
pass the first field (the container path) to basename. We also add a
list of attributes and a quick summary of each to the installer
configuration sample.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>