cubeit-installer: Consolidate the grub install
The grub install mechanism was not working properly for legacy + EFI,
and there are one too many places that it lives.

This patch consolidates the grub install logic to just call
grub_install in a single place.  The side effect is that the partition
label variables need to be the same all the scripts but the content of
the variables of course can vary.

Fixes beyond the consolidation of the logic:
 * Consistently use the disk label variables.
 * Remove the grub /images directory from the generated
   installer image in the EFI partition, which makes
   the installer and run time images the same with respect
   to the paths.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Reviewed-by: Mark Asselstine <mark.asselstine@windriver.com>