installer: make VRF container detection based on network attributes
Rather than keying off a specific container name, we change the installer
to lookup the VRF container via the "net=vrf" attribute.

We also move the "vrf" runtime attribute to the VRF container, and leave
dom0 to simply be our control domain.

We use the newly added "network-overlay" network type to assign the
VRF container 192.168.42.4, and allow the 'nslinkup' hook to configure
the IP at container creation time.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>