cubeit-installer: remove unneeded 'umount'
Commit 10544a30de35 [Use explicit, absolute paths for file operations]
changed the nested mount of 'mnt' on '${TMPMNT}/rootfs' to a nested
mount on '${BOOTTMPMNT}/mnt'. The associated umount of
'${BOOTTMPMNT}/mnt' takes place just above the if block we are
manipulating here. As such this umount is not needed and currently
shows up as a failure when you run create the install.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>