cubeit-installer: Use netprime name in the netprime setting section
Depending on what ever container runs last for the
HDD_INSTALL_CONTAINERS it will stay set as the cname in the netprime
section.  If the cube-vrf was installed last, it will partially get
treated like the netprime and the real netprime will not get the
system-resolved service turned off.  The simple fix is to just use the
correct variable.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>