cubeit: Add the 'app' container attribute
Up to now we have always used the default 'app' for our containers,
that being '/sbin/init'. Now we want to use tini for the cube-vrf and
possibly for other microservice containers in the future, which
requires that we pass args or possibly even use an app which is
different than '/sbin/init'. We therefor add a new 'app' attribute
which can be specified per container.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>