cubeit-installer: Fix up multi target grub-install
The verify-commands section is looking for the host grub-install on
the host, but ends up using that command for what is run in the chroot
on the end target image.  This needed to be separated into a
HOST_CMD_GRUB_INSTALL and a CMD_GRUB_INSTALL.  The CMD_GRUB_INSTALL is
specified in the config-usb.sh template.

Once separated from the CMD_GRUB_INSTALL, the host binaries can be
used to add either the efi or legacy bios grub files to the target
image.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>