cubeit-installer: keep systemd-getty-generator under control
The systemd-getty-generator will interpret the kernel cmdline
"console" to create serial-getty@XXX.service files automatically. This
will result in a battle between the screen and serial gettys and thus
an unusable terminal. By creating a link for the interface we know we
want to use with screen, ie. SCREEN_GETTY_CONSOLE%, to /dev/null we
effective stop systemd-getty-generator from creating the serial-getty
and messing things up.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>