housekeeping: be consistent with the use/non-use of the ':' in debugmsg
Sometimes we use ':' after [ERROR]/[INFO]/[WARNING] in debugmsg,
sometimes we don't. This comes across as sloppy and makes any
postprocessing of the logs a pain. Here we remove the ':' in all
cases, though I could have choosen to go the other way and add ':'
where missing, it seems neater to remove its use.

We still see some ':' in the output as cube-ctl and cube-cfg need to
be updated to be consistent as well.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>